Codette_final/requirements.txt
# Core AI Libraries
ollama  # Replacing transformers with Ollama for local Llama 3 execution
torch
faiss-cpu>=1.7.4
numpy>=1.23.0

# Security & Utilities
cryptography>=41.0.0
jwt>=1.3.1
python-dotenv>=1.0.0

# Speech & Voice Processing
pyttsx3>=2.90
speechrecognition>=3.10.0

# Async & Web API
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.24.0
asyncio>=3.4.3

# Vector Databases & AI Memory
pinecone-client>=2.2.4
chromadb>=0.4.5